requirements.txt
flask
llama-cpp-python
pandas
tabulate